Migrate listing test to contractstore
