+++ internal/contractgateway/smartcontractgw_test.go
@@ -820,132 +820,6 @@ func TestPostDeployMissingContractAddress(t *testing.T) {
 	assert.Regexp("message1: Missing contract address in receipt", err.Error())
 }
 
-func TestBuildIndex(t *testing.T) {
-	assert := assert.New(t)
-	dir := tempdir()
-	defer cleanup(dir)
-
-	// Migration of legacy contract interfaces
-
-	var emptySwagger spec.Swagger
-	swaggerBytes, _ := json.Marshal(&emptySwagger)
-	ioutil.WriteFile(path.Join(dir, "contract_0123456789abcdef0123456789abcdef01234567.swagger.json"), swaggerBytes, 0644)
-
-	okSwagger := spec.Swagger{
-		SwaggerProps: spec.SwaggerProps{
-			Info: &spec.Info{
-				InfoProps: spec.InfoProps{
-					Title: "good one",
-				},
-			},
-		},
-	}
-	okSwagger.Info.AddExtension("x-firefly-deployment-id", "840b629f-2e46-413b-9671-553a886ca7bb")
-	swaggerBytes, _ = json.Marshal(&okSwagger)
-	ioutil.WriteFile(path.Join(dir, "contract_123456789abcdef0123456789abcdef012345678.swagger.json"), swaggerBytes, 0644)
-
-	regSwagger := spec.Swagger{
-		SwaggerProps: spec.SwaggerProps{
-			Info: &spec.Info{
-				InfoProps: spec.InfoProps{
-					Title: "good one",
-				},
-			},
-		},
-	}
-	regSwagger.Info.AddExtension("x-firefly-deployment-id", "840b629f-2e46-413b-9671-553a886ca7bb")
-	regSwagger.Info.AddExtension("x-firefly-registered-name", "migratedcontract")
-	swaggerBytes, _ = json.Marshal(&regSwagger)
-	ioutil.WriteFile(path.Join(dir, "contract_23456789abcdef0123456789abcdef0123456789.swagger.json"), swaggerBytes, 0644)
-
-	ioutil.WriteFile(path.Join(dir, "contract_3456789abcdef0123456789abcdef01234567890.swagger.json"), []byte(":bad swagger"), 0644)
-
-	// New contract interfaces
-	info1 := &contractregistry.ContractInfo{
-		Address:      "456789abcdef0123456789abcdef012345678901",
-		ABI:          "840b629f-2e46-413b-9671-553a886ca7bb",
-		Path:         "/contracts/456789abcdef0123456789abcdef012345678901",
-		SwaggerURL:   "http://localhost:8080/contracts/456789abcdef0123456789abcdef012345678901?swagger",
-		RegisteredAs: "",
-		TimeSorted: messages.TimeSorted{
-			CreatedISO8601: time.Now().UTC().Format(time.RFC3339),
-		},
-	}
-	info1Bytes, _ := json.Marshal(info1)
-	ioutil.WriteFile(path.Join(dir, "contract_456789abcdef0123456789abcdef012345678901.instance.json"), info1Bytes, 0644)
-	info2 := &contractregistry.ContractInfo{
-		Address:      "56789abcdef0123456789abcdef0123456789012",
-		ABI:          "840b629f-2e46-413b-9671-553a886ca7bb",
-		Path:         "/contracts/somecontract",
-		SwaggerURL:   "http://localhost:8080/contracts/somecontract?swagger",
-		RegisteredAs: "somecontract",
-		TimeSorted: messages.TimeSorted{
-			CreatedISO8601: time.Now().UTC().Format(time.RFC3339),
-		},
-	}
-	info2Bytes, _ := json.Marshal(info2)
-	ioutil.WriteFile(path.Join(dir, "contract_56789abcdef0123456789abcdef0123456789012.instance.json"), info2Bytes, 0644)
-
-	deployMsg := &messages.DeployContract{
-		ContractName: "abideployable",
-	}
-	deployBytes, _ := json.Marshal(&deployMsg)
-	ioutil.WriteFile(path.Join(dir, "abi_840b629f-2e46-413b-9671-553a886ca7bb.deploy.json"), deployBytes, 0644)
-	ioutil.WriteFile(path.Join(dir, "abi_e27be4cf-6ae2-411e-8088-db2992618938.deploy.json"), deployBytes, 0644)
-	ioutil.WriteFile(path.Join(dir, "abi_519526b2-0879-41f4-93c0-09acaa62e2da.deploy.json"), []byte(":bad json"), 0644)
-
-	s, _ := NewSmartContractGateway(
-		&SmartContractGatewayConf{
-			StoragePath: dir,
-		},
-		&tx.TxnProcessorConf{
-			OrionPrivateAPIS: false,
-		},
-		nil, nil, nil, nil,
-	)
-	scgw := s.(*smartContractGW)
-
-	contracts := scgw.cs.ListContracts()
-	assert.Equal(4, len(contracts))
-	info, err := scgw.cs.GetContractByAddress("123456789abcdef0123456789abcdef012345678")
-	assert.NoError(err)
-	assert.Equal("123456789abcdef0123456789abcdef012345678", info.Address)
-
-	req := httptest.NewRequest("GET", "/contracts", bytes.NewReader([]byte{}))
-	res := httptest.NewRecorder()
-	params := httprouter.Params{}
-	scgw.listContractsOrABIs(res, req, params)
-	assert.Equal(200, res.Result().StatusCode)
-	var contractInfos []*contractregistry.ContractInfo
-	err = json.NewDecoder(res.Body).Decode(&contractInfos)
-	assert.NoError(err)
-	assert.Equal(4, len(contractInfos))
-	assert.Equal("123456789abcdef0123456789abcdef012345678", contractInfos[0].Address)
-	assert.Equal("23456789abcdef0123456789abcdef0123456789", contractInfos[1].Address)
-	assert.Equal("456789abcdef0123456789abcdef012345678901", contractInfos[2].Address)
-	assert.Equal("56789abcdef0123456789abcdef0123456789012", contractInfos[3].Address)
-
-	somecontractAddr, err := scgw.cs.ResolveContractAddress("somecontract")
-	assert.NoError(err)
-	assert.Equal("56789abcdef0123456789abcdef0123456789012", somecontractAddr)
-
-	migratedcontractAddr, err := scgw.cs.ResolveContractAddress("migratedcontract")
-	assert.NoError(err)
-	assert.Equal("23456789abcdef0123456789abcdef0123456789", migratedcontractAddr)
-
-	req = httptest.NewRequest("GET", "/abis", bytes.NewReader([]byte{}))
-	res = httptest.NewRecorder()
-	params = httprouter.Params{}
-	scgw.listContractsOrABIs(res, req, params)
-	assert.Equal(200, res.Result().StatusCode)
-	var abiInfos []*contractregistry.ABIInfo
-	err = json.NewDecoder(res.Body).Decode(&abiInfos)
-	assert.NoError(err)
-	assert.Equal(2, len(abiInfos))
-	assert.Equal("840b629f-2e46-413b-9671-553a886ca7bb", abiInfos[0].ID)
-	assert.Equal("e27be4cf-6ae2-411e-8088-db2992618938", abiInfos[1].ID)
-}
-
 func TestGetContractOrABIFail(t *testing.T) {
 	assert := assert.New(t)
 	dir := tempdir()
