+++ internal/contractregistry/contractstore_test.go
@@ -22,6 +22,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/go-openapi/spec"
 	"github.com/hyperledger-labs/firefly-ethconnect/internal/eth"
 	"github.com/hyperledger-labs/firefly-ethconnect/internal/messages"
 	"github.com/stretchr/testify/assert"
@@ -207,4 +208,106 @@ func TestCheckNameAvailableRRFail(t *testing.T) {
 
 	err := cs.CheckNameAvailable("lobster", true)
 	assert.EqualError(err, "pop")
+}
+
+func TestBuildIndex(t *testing.T) {
+	assert := assert.New(t)
+	dir := tempdir()
+	defer cleanup(dir)
+
+	// Migration of legacy contract interfaces
+
+	var emptySwagger spec.Swagger
+	swaggerBytes, _ := json.Marshal(&emptySwagger)
+	ioutil.WriteFile(path.Join(dir, "contract_0123456789abcdef0123456789abcdef01234567.swagger.json"), swaggerBytes, 0644)
+
+	okSwagger := spec.Swagger{
+		SwaggerProps: spec.SwaggerProps{
+			Info: &spec.Info{
+				InfoProps: spec.InfoProps{
+					Title: "good one",
+				},
+			},
+		},
+	}
+	okSwagger.Info.AddExtension("x-firefly-deployment-id", "840b629f-2e46-413b-9671-553a886ca7bb")
+	swaggerBytes, _ = json.Marshal(&okSwagger)
+	ioutil.WriteFile(path.Join(dir, "contract_123456789abcdef0123456789abcdef012345678.swagger.json"), swaggerBytes, 0644)
+
+	regSwagger := spec.Swagger{
+		SwaggerProps: spec.SwaggerProps{
+			Info: &spec.Info{
+				InfoProps: spec.InfoProps{
+					Title: "good one",
+				},
+			},
+		},
+	}
+	regSwagger.Info.AddExtension("x-firefly-deployment-id", "840b629f-2e46-413b-9671-553a886ca7bb")
+	regSwagger.Info.AddExtension("x-firefly-registered-name", "migratedcontract")
+	swaggerBytes, _ = json.Marshal(&regSwagger)
+	ioutil.WriteFile(path.Join(dir, "contract_23456789abcdef0123456789abcdef0123456789.swagger.json"), swaggerBytes, 0644)
+
+	ioutil.WriteFile(path.Join(dir, "contract_3456789abcdef0123456789abcdef01234567890.swagger.json"), []byte(":bad swagger"), 0644)
+
+	// New contract interfaces
+	info1 := &ContractInfo{
+		Address:      "456789abcdef0123456789abcdef012345678901",
+		ABI:          "840b629f-2e46-413b-9671-553a886ca7bb",
+		Path:         "/contracts/456789abcdef0123456789abcdef012345678901",
+		SwaggerURL:   "http://localhost:8080/contracts/456789abcdef0123456789abcdef012345678901?swagger",
+		RegisteredAs: "",
+		TimeSorted: messages.TimeSorted{
+			CreatedISO8601: time.Now().UTC().Format(time.RFC3339),
+		},
+	}
+	info1Bytes, _ := json.Marshal(info1)
+	ioutil.WriteFile(path.Join(dir, "contract_456789abcdef0123456789abcdef012345678901.instance.json"), info1Bytes, 0644)
+	info2 := &ContractInfo{
+		Address:      "56789abcdef0123456789abcdef0123456789012",
+		ABI:          "840b629f-2e46-413b-9671-553a886ca7bb",
+		Path:         "/contracts/somecontract",
+		SwaggerURL:   "http://localhost:8080/contracts/somecontract?swagger",
+		RegisteredAs: "somecontract",
+		TimeSorted: messages.TimeSorted{
+			CreatedISO8601: time.Now().UTC().Format(time.RFC3339),
+		},
+	}
+	info2Bytes, _ := json.Marshal(info2)
+	ioutil.WriteFile(path.Join(dir, "contract_56789abcdef0123456789abcdef0123456789012.instance.json"), info2Bytes, 0644)
+
+	deployMsg := &messages.DeployContract{
+		ContractName: "abideployable",
+	}
+	deployBytes, _ := json.Marshal(&deployMsg)
+	ioutil.WriteFile(path.Join(dir, "abi_840b629f-2e46-413b-9671-553a886ca7bb.deploy.json"), deployBytes, 0644)
+	ioutil.WriteFile(path.Join(dir, "abi_e27be4cf-6ae2-411e-8088-db2992618938.deploy.json"), deployBytes, 0644)
+	ioutil.WriteFile(path.Join(dir, "abi_519526b2-0879-41f4-93c0-09acaa62e2da.deploy.json"), []byte(":bad json"), 0644)
+
+	cs := NewContractStore("", dir, nil)
+	cs.Init()
+
+	contracts := cs.ListContracts()
+	assert.Equal(4, len(contracts))
+	assert.Equal("123456789abcdef0123456789abcdef012345678", contracts[0].(*ContractInfo).Address)
+	assert.Equal("23456789abcdef0123456789abcdef0123456789", contracts[1].(*ContractInfo).Address)
+	assert.Equal("456789abcdef0123456789abcdef012345678901", contracts[2].(*ContractInfo).Address)
+	assert.Equal("56789abcdef0123456789abcdef0123456789012", contracts[3].(*ContractInfo).Address)
+
+	info, err := cs.GetContractByAddress("123456789abcdef0123456789abcdef012345678")
+	assert.NoError(err)
+	assert.Equal("123456789abcdef0123456789abcdef012345678", info.Address)
+
+	somecontractAddr, err := cs.ResolveContractAddress("somecontract")
+	assert.NoError(err)
+	assert.Equal("56789abcdef0123456789abcdef0123456789012", somecontractAddr)
+
+	migratedcontractAddr, err := cs.ResolveContractAddress("migratedcontract")
+	assert.NoError(err)
+	assert.Equal("23456789abcdef0123456789abcdef0123456789", migratedcontractAddr)
+
+	abis := cs.ListABIs()
+	assert.Equal(2, len(abis))
+	assert.Equal("840b629f-2e46-413b-9671-553a886ca7bb", abis[0].(*ABIInfo).ID)
+	assert.Equal("e27be4cf-6ae2-411e-8088-db2992618938", abis[1].(*ABIInfo).ID)
 }
